+++ src/index.ts
@@ -17,6 +17,8 @@ export { default as RadioButton } from '@/package/components/radio-button.vue';
 export { default as Tabs } from '@/package/components/tabs.vue';
 export { default as Tab } from '@/package/components/tab.vue';
 export { default as TabPanel } from '@/package/components/tab-panel.vue';
+export { default as Accordion } from '@/package/components/accordion.vue';
+export { default as AccordionItem } from '@/package/components/accordion-item.vue';
 export { default as Toast } from '@/package/components/toast.vue';
 export { default as ToastContainer } from '@/package/components/toast-container.vue';
 export { default as Alert } from '@/package/components/alert.vue';